+++ feature/src/main/java/com/gdg/feature/map/MapRoute.kt
@@ -104,7 +104,7 @@ fun MapRoute(
         locations = mapViewModel.locations,
         roads = roads,
         congestionState = getCongestionState,
-        getPlaceEntity = { id -> mapViewModel.getCongestion(id) },
+        getPlaceEntity = { id -> mapViewModel.getCongestion(id.toInt()) },
         onButtonClick = mapViewModel::navigateToDetail
     )
 }
